Rename cadl-ranch references to spector and update terminology
- Rename gen-cadl-ranch.js to gen-spector.js
- Rename cadl-ranch-list.js to spector-list.js
- Update all references in package.json scripts, launch.json, CI pipelines,
  CONTRIBUTING.md, copilot-instructions.md, and SKILL.md
- Update spector-list.js with latest entries from upstream
- Replace HLC terminology with AutoRest in CI and test docs
- Update README.md option descriptions to remove Modular jargon

Fixes Azure/autorest.typescript#3792
