+++ packages/typespec-ts/package.json
@@ -45,10 +45,10 @@
     "generate-and-run:modular": "npm run generate-tsp-only:modular && npm run integration-test:alone:modular && npm run stop-test-server -- -p 3002",
     "generate-and-run:azure-modular": "npm run generate-tsp-only:azure-modular && npm run integration-test:alone:azure-modular && npm run stop-test-server -- -p 3002",
     "generate-tsp-only": "npm run generate-tsp-only:azure-rlc && npm run generate-tsp-only:rlc &&npm run generate-tsp-only:azure-modular && npm run generate-tsp-only:modular",
-    "generate-tsp-only:rlc": "npx tsx ./test/commands/gen-cadl-ranch.js --tag=rlc",
-    "generate-tsp-only:azure-rlc": "npx tsx ./test/commands/gen-cadl-ranch.js --tag=azure-rlc",
-    "generate-tsp-only:modular": "npx tsx ./test/commands/gen-cadl-ranch.js --tag=modular",
-    "generate-tsp-only:azure-modular": "npx tsx ./test/commands/gen-cadl-ranch.js --tag=azure-modular",
+    "generate-tsp-only:rlc": "npx tsx ./test/commands/gen-spector.js --tag=rlc",
+    "generate-tsp-only:azure-rlc": "npx tsx ./test/commands/gen-spector.js --tag=azure-rlc",
+    "generate-tsp-only:modular": "npx tsx ./test/commands/gen-spector.js --tag=modular",
+    "generate-tsp-only:azure-modular": "npx tsx ./test/commands/gen-spector.js --tag=azure-modular",
     "regen-test-baselines": "npm run generate-tsp-only",
     "integration-test:alone": "npm run integration-test:alone:rlc && npm run integration-test:alone:azure-rlc && npm run integration-test:alone:modular && npm run integration-test:alone:azure-modular",
     "integration-test:alone:rlc": "vitest run --project integration-rlc",
