+++ packages/typespec-ts/test/azureModularIntegration/payloadMultipart.spec.ts
@@ -23,7 +23,7 @@
 //       id: "123",
 //       profileImage: {
 //         contents: fs.createReadStream(imgPath),
-//         // must specify a filename due to cadl-ranch limitations
+//         // must specify a filename due to spector limitations
 //         filename: "test.jpg"
 //       }
 //     });
