+++ packages/typespec-ts/test/commands/spector-list.js
