+++ packages/typespec-ts/test/modularUnit/scenarios/operations/wrapNonModelReturn.md
@@ -1,6 +1,6 @@
 # wrap-non-model-return wraps string scalar response with body property
 
-When `wrap-non-model-return` is enabled, scalar/non-model responses are wrapped with a `body` property to maintain HLC compatibility.
+When `wrap-non-model-return` is enabled, scalar/non-model responses are wrapped with a `body` property to maintain AutoRest backward compatibility.
 
 ## TypeSpec
 
@@ -281,7 +281,7 @@ export async function get(
 
 # wrap-non-model-return does not wrap array-of-models response
 
-Array-of-models responses (Composite kind) are not wrapped — they are returned as `Resource[]` directly (matching HLC behavior where PropertyKind.Composite means no body wrapper).
+Array-of-models responses (Composite kind) are not wrapped — they are returned as `Resource[]` directly (matching AutoRest behavior where PropertyKind.Composite means no body wrapper).
 
 ## TypeSpec
 
@@ -854,7 +854,7 @@ export async function getString(
 # wrap-non-model-return wraps LRO string response with body property
 
 When `wrap-non-model-return` is enabled, LRO operations with non-model final result types
-(e.g., string) should also be wrapped in a response type alias for HLC compatibility.
+(e.g., string) should also be wrapped in a response type alias for AutoRest backward compatibility.
 
 ## TypeSpec
 
@@ -952,7 +952,7 @@ export function getIkeSas(
 
 When both `wrap-non-model-return` and `head-as-boolean` are enabled and the operation is a HEAD request with no response body,
 the response is wrapped as `{ body: boolean }` to indicate resource existence (true = exists, false = not found).
-This matches HLC behavior for HEAD operations.
+This matches AutoRest behavior for HEAD operations.
 
 ## TypeSpec
